scraper/requirements.txt
pip
requests
unidecode
xmltodict
Scrapy==1.8.0
Scrapy-UserAgents
twisted==20.3.0
redis==3.4.1
google-cloud-storage==1.26.0
jsonschema==3.0.2
python-dateutil==2.8.0
google-cloud-firestore
pandas